+++ app/models/student_loans/slug_sequence.rb
@@ -32,6 +32,9 @@ module StudentLoans
       "student-loan-amount",
       "email-address",
       "email-verification",
+      "provide-mobile-number",
+      "mobile-number",
+      "mobile-verification",
       "bank-or-building-society",
       "personal-bank-account",
       "building-society-account",
@@ -54,6 +57,8 @@ module StudentLoans
         sequence.delete("student-loan-start-date") if claim.no_student_loan? || claim.student_loan_country_with_one_plan?
         sequence.delete("personal-bank-account") if claim.bank_or_building_society == "building_society"
         sequence.delete("building-society-account") if claim.bank_or_building_society == "personal_bank_account"
+        sequence.delete("mobile-number") if claim.provide_mobile_number == false
+        sequence.delete("mobile-verification") if claim.provide_mobile_number == false
         sequence.delete("ineligible") unless claim.eligibility.ineligible?
       end
     end